# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PetChain API",
    "version": "1.0.0",
...
    "contact": {
      "name": "PetChain Engineering",
      "email": "engineering@petchain.app",
      "url": "https://petchain.app"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    },
    {
      "url": "https://staging.petchain.app/api",
      "description": "Staging environment"
    },
    {
      "url": "https://api.petchain.app/api",
      "description": "Production environment"
    }
  ],
...
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List all users",
        "description": "Returns a paginated list of users. Requires admin role.",
        "operationId": "listUsers",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          },
          {
            "$ref": "#/components/parameters/SortByParam"
          },
          {
            "$ref": "#/components/parameters/SortOrderParam"
          },
          {
            "name": "role",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["owner", "vet", "admin"]
            },
            "description": "Filter by user role"
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search by name or email (case-insensitive)",
            "example": "jane"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of users",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/User"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "tags": ["Users"],
        "summary": "Get current authenticated user",
        "description": "Returns the profile of the currently authenticated user.",
        "operationId": "getCurrentUser",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user profile",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user by ID",
        "description": "Returns a user profile by ID. Admins can access any user; regular users can only access their own profile.",
        "operationId": "getUserById",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update user profile",
        "description": "Updates editable user profile fields. Admins can update any user; regular users can only update their own profile.",
        "operationId": "updateUser",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              },
              "example": {
                "name": "Jane Smith",
                "phone": "+14155552671"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UpdateUserResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user account",
        "description": "Permanently deletes a user account. Requires admin role.",
        "operationId": "deleteUser",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdParam"
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully (no content)"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/pets": {
      "get": {
        "tags": ["Pets"],
        "summary": "List pets",
        "description": "Returns a paginated list of pets. Owners see only their own pets; vets and admins can see all pets.",
        "operationId": "listPets",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          },
          {
            "$ref": "#/components/parameters/SortByParam"
          },
          {
            "$ref": "#/components/parameters/SortOrderParam"
          },
          {
            "name": "ownerId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter pets by owner ID",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          {
            "name": "species",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["dog", "cat", "bird", "rabbit", "other"]
            },
            "description": "Filter by species"
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search by pet name (case-insensitive)",
            "example": "buddy"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of pets",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Pet"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Pets"],
        "summary": "Create a new pet",
        "description": "Registers a new pet. The `ownerId` must match the authenticated user's ID (or the user must be an admin).",
        "operationId": "createPet",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePetRequest"
              },
              "example": {
                "name": "Buddy",
                "species": "dog",
                "breed": "Golden Retriever",
                "dateOfBirth": "2021-06-15",
                "microchipId": "985141002512345",
                "ownerId": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pet created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Pet"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from `POST /auth/login` or `POST /auth/register`.\n\n**Example:**\n```\nAuthorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\n```\n\nTokens expire after 7 days. Use `POST /auth/refresh` to obtain a new token."
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for server-to-server integrations. Contact engineering@petchain.app to obtain a key."
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "required": ["success", "data", "timestamp"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {},
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00Z"
          }
        }
      },
      "ApiError": {
        "type": "object",
        "required": ["success", "error", "timestamp"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string",
                "example": "Invalid input data"
              },
              "details": {
                "type": "object",
                "additionalProperties": true,
                "example": {
                  "field": "email",
                  "issue": "Invalid format"
                }
              }
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00Z"
          }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "required": ["page", "limit", "total", "totalPages", "hasNext", "hasPrev"],
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1,
            "example": 1
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100,
            "example": 20
          },
          "total": {
            "type": "integer",
            "minimum": 0,
            "example": 42
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0,
            "example": 3
          },
          "hasNext": {
            "type": "boolean",
            "example": true
          },
          "hasPrev": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "PaginatedResponse": {
        "type": "object",
        "required": ["success", "data", "pagination", "timestamp"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {}
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00Z"
          }
        }
      },
      "UUID": {
        "type": "string",
        "format": "uuid",
        "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
      },
      "ISODate": {
        "type": "string",
        "format": "date",
        "example": "2024-01-15"
      },
      "ISODateTime": {
        "type": "string",
        "format": "date-time",
        "example": "2024-01-15T10:30:00Z"
      },
...
      "User": {
        "type": "object",
        "required": [
          "id",
          "email",
          "name",
          "role",
          "pets",
          "createdAt",
          "updatedAt",
          "isEmailVerified",
          "authProvider"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/UUID"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "jane.doe@example.com"
          },
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "phone": {
            "type": "string",
            "example": "+14155552671"
          },
          "role": {
            "type": "string",
            "enum": ["owner", "vet", "admin"],
            "example": "owner",
            "description": "User role controlling access permissions"
          },
          "authProvider": {
            "type": "string",
            "enum": ["local", "google", "apple"],
            "example": "local"
          },
          "isEmailVerified": {
            "type": "boolean",
            "example": true
          },
          "lastLoginAt": {
            "$ref": "#/components/schemas/ISODateTime"
          },
          "pets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PetReference"
            },
            "description": "Lightweight references to pets owned by this user"
          },
          "createdAt": {
            "$ref": "#/components/schemas/ISODateTime"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/ISODateTime"
          }
        }
      },
      "PetReference": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/UUID"
          },
          "name": {
            "type": "string",
            "example": "Buddy"
          }
        }
      },
      "UpdateUserRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "example": "Jane Smith"
          },
          "phone": {
            "type": "string",
            "example": "+14155552671"
          },
          "role": {
            "type": "string",
            "enum": ["owner", "vet", "admin"],
            "example": "vet"
          }
        }
      },
      "UpdateUserResponse": {
        "type": "object",
        "required": ["id", "email", "name", "role", "updatedAt"],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/UUID"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "jane.doe@example.com"
          },
          "name": {
            "type": "string",
            "example": "Jane Smith"
          },
          "phone": {
            "type": "string",
            "example": "+14155552671"
          },
          "role": {
            "type": "string",
            "enum": ["owner", "vet", "admin"],
            "example": "vet"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/ISODateTime"
          }
        }
      },
      "Pet": {
        "type": "object",
        "required": ["id", "name", "species", "ownerId", "createdAt", "updatedAt"],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/UUID"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "Buddy"
          },
          "species": {
            "type": "string",
            "enum": ["dog", "cat", "bird", "rabbit", "other"],
            "example": "dog"
          },
          "breed": {
            "type": "string",
            "example": "Golden Retriever"
          },
          "dateOfBirth": {
            "$ref": "#/components/schemas/ISODate"
          },
          "microchipId": {
            "type": "string",
            "example": "985141002512345",
            "description": "15-digit ISO 11784/11785 microchip identifier"
          },
          "photoUrl": {
            "type": "string",
            "format": "uri",
            "example": "https://storage.petchain.app/pets/buddy.jpg"
          },
          "ownerId": {
            "$ref": "#/components/schemas/UUID"
          },
          "owner": {
            "type": "object",
            "properties": {
              "id": {
                "$ref": "#/components/schemas/UUID"
              },
              "name": {
                "type": "string",
                "example": "Jane Doe"
              },
              "email": {
                "type": "string",
                "format": "email",
                "example": "jane.doe@example.com"
              }
            }
          },
          "createdAt": {
            "$ref": "#/components/schemas/ISODateTime"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/ISODateTime"
          }
        }
      },
      "CreatePetRequest": {
        "type": "object",
        "required": ["name", "species", "ownerId"],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "Buddy"
          },
          "species": {
            "type": "string",
            "enum": ["dog", "cat", "bird", "rabbit", "other"],
            "example": "dog"
          },
          "breed": {
            "type": "string",
            "example": "Golden Retriever"
          },
          "dateOfBirth": {
            "$ref": "#/components/schemas/ISODate",
            "description": "Pet date of birth in YYYY-MM-DD format"
          },
          "microchipId": {
            "type": "string",
            "example": "985141002512345"
          },
          "photoUrl": {
            "type": "string",
            "format": "uri",
            "example": "https://storage.petchain.app/pets/buddy.jpg"
          },
          "ownerId": {
            "$ref": "#/components/schemas/UUID"
          }
        }
      },
...
    },
    "parameters": {
      "PageParam": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        },
        "description": "Page number for pagination"
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        },
        "description": "Number of items per page"
      },
      "SortByParam": {
        "name": "sortBy",
        "in": "query",
        "schema": {
          "type": "string"
        },
        "description": "Field to sort by (e.g. \"createdAt\", \"name\")"
      },
      "SortOrderParam": {
        "name": "sortOrder",
        "in": "query",
        "schema": {
          "type": "string",
          "enum": ["asc", "desc"],
          "default": "desc"
        },
        "description": "Sort direction"
      },
      "UserIdParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "User ID (UUID)",
        "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request — invalid input or missing required fields",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "success": false,
              "error": {
                "code": "VALIDATION_ERROR",
                "message": "Invalid input data",
                "details": {
                  "field": "email",
                  "issue": "Invalid email format"
                }
              },
              "timestamp": "2024-01-15T10:30:00Z"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized — missing or invalid JWT token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "success": false,
              "error": {
                "code": "UNAUTHORIZED",
                "message": "Authentication required. Please log in."
              },
              "timestamp": "2024-01-15T10:30:00Z"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden — authenticated but insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "success": false,
              "error": {
                "code": "FORBIDDEN",
                "message": "Access denied."
              },
              "timestamp": "2024-01-15T10:30:00Z"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "success": false,
              "error": {
                "code": "NOT_FOUND",
                "message": "Resource not found."
              },
              "timestamp": "2024-01-15T10:30:00Z"
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "success": false,
              "error": {
                "code": "INTERNAL_ERROR",
                "message": "Server error. Please try again later."
              },
              "timestamp": "2024-01-15T10:30:00Z"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}